pyproject.toml
[project]
name = "hybrid-ai-agent"
version = "0.1.0"
description = "Hybrid AI Coding Agent with LangGraph, Ollama, and Claude"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.3.0,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "httpx>=0.27.0,<0.28.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "rich>=13.0.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "ruff>=0.5.0,<0.6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "S", "B"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S108"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
